Updated earlier labs

+++ lab5_2.py
@@ -1,12 +1,15 @@
 import numpy as np
+
+
 def main():
 	n = 7
-	arr=np.zeros((n,n))
+	arr = np.zeros((n, n), dtype=np.int8)
 
 	for i in range(n):
 		for j in range(n - i - 1, n):
 			arr[i][j] = j - (n - i - 1) + 1
 	print(arr)
+
 
 if __name__ == '__main__':
 	main()
